+++ dynamite/parameter_space.py
@@ -671,7 +671,7 @@ class LegacyGridSearch(ParameterGenerator):
                     # Explicitly set minstep=0 to allow arbitrarily
                     # small steps, not recommended.
                     self.minstep.append(settings['minstep'] \
-                        if 'minstep' in settings else self.step)
+                        if 'minstep' in settings else settings['step'])
                 else:
                     self.step.append(None)
                     self.minstep.append(None)
@@ -745,7 +745,7 @@ class LegacyGridSearch(ParameterGenerator):
                     for s in [-1, 1]:
                         new_raw_value = np.clip(raw_center + s*step, lo, hi)
                         if abs(new_raw_value-par.raw_value) \
-                           > minstep - sys.float_info.epsilon:
+                           >= minstep - sys.float_info.epsilon:
                             self.new_parset[paridx].raw_value = new_raw_value
                             if self._is_newmodel(self.new_parset, eps=1e-10):
                                 self.model_list.append\
@@ -827,7 +827,7 @@ class GridWalk(ParameterGenerator):
                     self.step.append(settings['step'])
                     # use 'minstep' value if present, otherwise use 'step'
                     self.minstep.append(settings['minstep'] \
-                        if 'minstep' in settings else self.step)
+                        if 'minstep' in settings else settings['step'])
                 else:
                     self.step.append(None)
                     self.minstep.append(None)
@@ -937,7 +937,7 @@ class GridWalk(ParameterGenerator):
             raw_values = []
             # start with lo...
             delta = center[paridx] - self.clip(center[paridx] - step, lo, hi)
-            if abs(delta) > minstep - sys.float_info.epsilon:
+            if abs(delta) >= minstep - sys.float_info.epsilon:
                 raw_values.append(self.clip(center[paridx] - step, lo, hi))
             # now mid... tol(erance) is necessary in case minstep < eps
             if len(raw_values) > 0:
@@ -951,7 +951,7 @@ class GridWalk(ParameterGenerator):
                 raw_values.append(self.clip(center[paridx], lo, hi))
             # and now hi...
             delta = self.clip(center[paridx] + step, lo, hi) - center[paridx]
-            if abs(delta) > minstep - sys.float_info.epsilon:
+            if abs(delta) >= minstep - sys.float_info.epsilon:
                 tol = abs(self.clip(center[paridx]+step,lo,hi)-raw_values[-1])
                 if abs(raw_values[-1]) > eps:
                     tol /= abs(raw_values[-1])
@@ -1051,7 +1051,7 @@ class FullGrid(ParameterGenerator):
                     self.step.append(settings['step'])
                     # use 'minstep' value if present, otherwise use 'step'
                     self.minstep.append(settings['minstep'] \
-                        if 'minstep' in settings else self.step)
+                        if 'minstep' in settings else settings['step'])
                 else:
                     self.step.append(None)
                     self.minstep.append(None)
@@ -1169,8 +1169,8 @@ class FullGrid(ParameterGenerator):
             # start with lo...
             while raw_value >= lo:
                 raw_new = self.clip(raw_value-step, lo, hi)
-                if abs(raw_value-raw_new) > max(minstep,eps) \
-                                            - sys.float_info.epsilon:
+                if abs(raw_value-raw_new) >= max(minstep,eps) \
+                                             - sys.float_info.epsilon:
                     raw_values.append(raw_new)
                 else:
                     break
@@ -1179,8 +1179,8 @@ class FullGrid(ParameterGenerator):
             raw_value = center[paridx]
             while raw_value <= hi:
                 raw_new = self.clip(raw_value+step, lo, hi)
-                if abs(raw_value-raw_new) > max(minstep,eps) \
-                                            - sys.float_info.epsilon:
+                if abs(raw_value-raw_new) >= max(minstep,eps) \
+                                             - sys.float_info.epsilon:
                     raw_values.append(raw_new)
                 else:
                     break